Allow selecting file answer type if file upload enabled for group
If file_upload_enabled is set to true for the group, show "File" as an
option when creating a question.

Have conditional validation for the page to select the answer type to
ensure that nobody can manipule the page HTML to create a question
with "File" answer type.

We don't need to do conditional validation when the `Page` is created
from the `DraftQuestion` as there is no way for users to manipulate
this.