Use PageRepository in specs where appropriate
We want to mock repositories instead of forms-api requests in our specs
as much as possible; this will make it easier to change the source of
the model data in future.

This commit changes specs that were mocking HTTP requests to
`/api/v1/forms/:form_id/pages/:page_id` so that they mock the
corresponding PageRepository methods instead.