Remove unused hint text
This was supposed to be shown when deleting a form that is live, but due
to a typo (there is no method `Forms#live?` in forms-admin), it never
did get shown. However, it's also inaccurate, because currently we can't
delete just a draft, and deleting a live form does delete the whole
form.

I'm not sure when we'll be implementing deleting a draft form, so for
now let's just remove this content. When we do get around to making it
possible to delete a draft form the content may change anyway, but if it
doesn't we can always put the code back.