Add a test for GDS Transport font loading
It's possible for changes to our build to cause assets not to be built
properly. One of the problems this causes is a failure to load the GDS
Transport font.

This commit adds a test which will fail if the GDS Transport font is
referenced in a font-face declaration but fails to load.