Use Page#routing_conditions instead of Page#conditions
Now that Page#routing_conditions returns Condition objects as expected
we can use that instead of Page#conditions; and then maybe in the future
we can use that method for something else.

This commit replaces all the uses of Page#condition I found using the
deprecation warning from the previous commit.