Refactor page delete confirmation view
Move logic around routing from the view to the controller for the
page#delete action.

The main aim is to reduce the repetition of the notification banner
call in the view; to achieve this we rely on the consistency of the
content, where the warning refers to up to two different questions; the
question about to be deleted, and the question that "owns" the route
(which depends on the kind of route involved).

A future refactor might appropriately move this logic out of the
controller to a service, or maybe something more generally useful like a
model that expresses the relationships between conditions and routes,
but before that we need to make this view work with cases where the page
to be deleted is associated with more than two routes, so for now I
think this is a good place to stop.