Add task for deleting a user
Adds a new task for deleting an individual user, and a dry run task for
checking it in advance.

Doesn't delete any associated forms or groups - this complicates things
quite a lot and it's less risky for live forms if we delete or reassign
those form/group associations prior to running this task to delete the
user.