Fix view spec for pages/type_of_answer
This spec was originally mocking a method that isn't called when the
view is rendered in production; when `url` is `nil` for `form_with`
ActionView tries to guess what the path for the form action should be
based on the class name of the model(s) passed in. However, when this
view is rendered the instance variable with the url should always be
set, because there is no route with a matching name in our app.

This commit changes the spec to set the instance variable appropriately.