Fix failure message for matcher #have_been_updated_to
When this matcher failed it was printing the original record as the
expected response, when the expected response is actually the updated
record. This commit fixes this.

This commit also tweaks the formatting of the failure message slightly
so that the array of requests is formatted in the same way as the
expected response, to make visual comparison slightly easier.