Send request headers when retrieving form documents from API
We were not sending the request headers containing the API key when
making a request to forms-api to retrieve form documents for creating
the CSV reports.

Also send an "Accept" header.

Assert in the tests that the expected headers are sent.

Set the forms_api.auth_key setting in test.yml so we can validate the
expected header is sent in the tests.