Refactor controllers to utilise form repository
This alters the controllers so that they are using the form_repository to access any form objects, rather than directly through the ActiveResource model.

For testing purposes, factory built forms that have not been mocked out through ActiveResource will not behave the same, meaning it is necessary to pass the form.id into a path method, rather than just the form object. This does not affect the actual operation of the app, just the tests.