Add task to delete users with no name or org set
We want to delete users accounts that haven't been used since we moved
over from having trial accounts to having trial groups, so we can delete
some old code.

To do this we've decided to start with deleting users that haven't got
an organisation or a name set. This means that they haven't signed in
since we started asking users for their name and organisation and they
weren't one of the trial users manually onboarded by the team.

This logic does also catch users who have signed in, seen that they were
asked for a name or organisation, and decided not to proceed, but we
think this is okay because they won't have lost any forms if we delete
them at that point.

To avoid a possible race condition where a user could have just signed
in for the first time as this task is run, we check whether the user
last signed in less than the time it would take for their session cookie
to expire.

This code is largely based on the spike in #1627.

Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>