Allow a maximum of 5 file upload questions
Display a validation error if a form creator attempts to add a 6th
file upload question.

There is a 40GB limit for sending emails using SES version 2. We
currently allow uploaded files to be a maximum of 7GB. Allowing 5
questions means that there will be a maximum of 35GB of uploaded
files. We need the total size of all files to be below 40GB, as the
content of the email will take up some of this size limit.