Try to maintain referential integrity when deleting users
Change users:delete_users_with_no_name_or_org to skip deletion if a form
cannot be deleted.

Initially I only handled the false case to keep RuboCop happy, it's
unlikely to happen unless there has been a programming error or
something else fundamentally wrong.

When asked about it I first thought it would be easier to keep around
the form but delete the user anyway, as that would be quicker to clean
up, but that does mean there is a broken reference in the forms-api
database, which isn't ideal.

If there is an issue deleting a form it's likely something fundamentally
wrong anyway, in which case it's probably better to fail fast.

It's a shame that we can't do something like a transaction with the
forms resources, where we only delete everything if everything can be
deleted, but the microservices architecture gets in the way of that.