Update question text error messages for file upload
For file upload questions, we have changed the label for the question
text input on the edit question page to be "Ask for a file". As a
result, we also need to update the error messages if the input is
blank or contains too many characters.

To accomplish this, add custom validation methods and use these
instead of the `QuestionTextValidation` concern. Define a shared
constant for the maximum length so that this is common between all the
places we validate the question text length.