Add branching_enabled to pages controller
To keep track of whether the advanced branching feature is enabled for
the current form, or if it's enabled for the whole environment, a new
method is added to the page controller.

It's a shame add more methods to the controller. Doing so allows us to
access value in all the places we need it.