Change redirect when making new route
When a user is creating a new route, they should go to the
current new route page.

If the page already has a primary route, they should be redirected to
the page for creating a secondary skip route.

At the moment, this is behind the branc_routing feature flag.

Without this flag, skip routes don't exist so we don't redirect to the
skip condition page unless the flag is enabled.