Add enabled_by_group? to feature_service
Add a new way to specify that feature flags should be per group
instead of set by environment variable.

To allow individual groups to have a feature flag, you need to do two
things.

Add the feature flag to the settings.yml file. For example, to add a
feature flag called branch_routing, follow these steps:

Add method to Group, called `branch_routing_enabled?`. You probably want
to add a new column to Group called `branch_routing_enabled` and rely on
the rails helper to add the `?` at the end. It should probably have a
default value of false.

Add a rake task to set/clear this value.

Add the following to the settings yaml config.

```yaml
features:
  branch_routing:
    enabled_by_group: true
```

Use FeatureService.new(group:
current_form.group).enabled?(:branch_routing) whenever you need to test
the feature flag.

The setting can be enabled for a whole environment, for example dev, in
the usual way:

```yaml
features:
  branch_routing: true
```

This commit uses:

```
return group.send(:"#{feature_name}_enabled?")
```

in FeatureService#enabled? to check if the group has the feature. It
will raise an exception if this method is not defined, so the changes to
group should be made before adding the feature to the settings file.

Raising an exception rather than returning false was chosen to expose
issues earlier in development.