Change FeatureService.new to use keywords
The FeatureService class can be initialized with single positional
argument to hold the user.

The user argument is optional and most features don't use it.

Adding other arguments in the future would require callers to provide
nil as a placeholder.

Instead the initializer has been changed to take a keyword argument for
user.

All the callers have been updated to use this new format.

This should make it easier to add new keywords in the future.