Log Auth0 session_id for the callback request after login
Log the session_id that we get in the post-login callback request from
Auth0 so that we can link this up to the login events logged by Auth0
itself. Without this, we have no link between the Auth0 logs and our
application logs meaning that if Auth0 returns a success login
response but we fail to handle it for any reason on our end, we can't
match up the 2 sets of logs.