Change blank error message when adding route
When the form creator adds a new route but doesn't select an option,
they are shown an error message asking them to do so.

This commit adds a new error message for when the branch_routing feature
is enabled.

It's a bit awkward to add validation error message which changes based
on a feature flag.

I chose a custom validation method. It could have also been done by
inheritance and changing the class used in the controller.

I couldn't get a dynamic message using a proc or lambda using the
standard presence validator.