Change content for adding routes
When the form creator uses the add new route button they should see
different content if the branch_routing feature is enabled.

The pages and translations have been refactored to make this a bit
clearer.

I think avoiding dynamic translations is probably a good idea, but
hopefully they won't be around for long.

I think further refactoring could remove the use of a policy to decide
if more routes could be added.