Add warning when user is about to delete condition with secondary skip
Currently (when branching is enabled) deleting a skip route for a
question that has a secondary skip route puts the form in a state where
there is a route with an error that can’t be deleted.

We do want users to be able to delete individual routes for a question.

So we’ve agreed that we will delete the secondary skip for a question
when deleting any other routes for that question. And we’ll let the user
know before the delete a route, with a warning.

This commit adds the warning; note that the deletion will take place in
forms-api, see alphagov/forms-api#686. To enable this we add a method
`#has_secondary_skip?` to the DeleteConfirmationInput object... I'm not
sure if this is the best place for this, but it'll do for now.