Add failing test for delete all routes
We've found that in production the "delete all routes" journey has
stopped working properly; after the user confirms that they want to
delete all routes for a question, they see an error page rather than
being redirected to the "add and edit questions" page as expected.

Debugging locally revealed this is an unintended side-effect of an
unrelated change in forms-api, in PR alphagov/forms-api#686, which
changes the conditions model to delete secondary skip routes when a
skip route is deleted.

Because the process to delete all routes deletes the primary condition
before the secondary skip, when it comes to delete the secondary skip
that no longer exists, and the API responds with a 404. This 404 is then
parlayed by ActiveResource into a ResourceNotFound exception, which then
bubbles up to the Rails middleware, and is treated as a record not found
error, and Rails renders a 404 error page.

This commit adds a test that simulates this interaction, so we can make
some changes in our system to handle the scenario better.