.review_apps: ensure container image URI is unique to every run
We want to make sure that the tag applied to the container being built in the
workflow is unique on each run, so that a re-run of a failed workflow will
result in a new container. There are many reasons for failure, and sometimes we
want the container to be built afresh.