Change page list component to use condition check page as route page
For "normal" skip routes, the check page and the routing page for the
condition are the same, but for the "any other answer" route in a branch
the secondary skip condition has the check page as the page that starts
the branch, which in the current model for branch routing is the page
that is of interest when editing the routes for the branch.

This commit adds tests describing the behaviour we expect in terms of
what the content should say, and changes the page list component to use
the condition check page in all places to make those tests pass.

It's probably a smell that we have to do this bookkeeping to model
routes, and maybe we should have a record that is a route.