Change the cannot route to next page error message for any other answer route specifically
We want to use different content for the error when a route has its
goto page before the page it is routing from if the route is an any
other answer route.

This commit adds the new content and tweaks the logic of the code to
generate the error messages for routing warnings to first look for a
specific any other answer error message before using the more generic
error message. This lets make other error message more specific in
future more easily, while keeping duplication to a minimum.

We also move the errors messages to the `errors.*` scope, so that
i18n-tasks doesn't complain about unused keys.