Refactor MailchimpListSynchronizer class
The current MailchimpListSynchronizer#synchronize method takes an array
of email addresses to update the mailchimp list.

It uses sets to decide which emails need to be added to mailchimp and
which need to be removed.

Because we plan to add other information to the mailchimp lists, such as
role, the method needs to change.

This refactor introduces a MailchimpMember class to hold the email
address and status.

This allows it to be extended with other attributes in the future.

The logic for updating and removing members has been kept but will allow
updating records where attributes have changed in the future.