Improve error logging in MailchimpListSynchronizer
When Mailchimp cannot complete a request, an exception is thrown and the
error is logged. The .title .detail attributes are logged after the
subscriber_hash and list_id.

These attributes aren't set on the exception returned.

Looking at the Mailchimp docs and API source, it looks like they are
contained in the response_body attribute, which should be a JSON object.

This commit changes the logging to directly log status and response_body
and combines these with the original error message to make them easier
to find in logs.

In the future, it may be better to extract the original attributes from the
response_body. For now the whole object is logged so we can see whats in
it.

See the following:
https://mailchimp.com/developer/marketing/docs/errors/#the-basics
https://github.com/mailchimp/mailchimp-marketing-ruby/blob/dd4fd7adeb10cae5afbb0a4e8c332ae23345bb8b/lib/MailchimpMarketing/api_client.rb#L83