Change Mailchimp logging to avoid errors
Mailchimp logging throws an error because the @response_body on the
ApiError is not a public method.

This commit fixes the logging for Mailchimp errors.

When the mailchimp API returns an error, the exception contains
information within the @response_body instance variable which isn't
exposed.

The changes to log_mailchimp_error extract this information when it's
valid and available and logs it as JSON.

It also takes care to filter out email addresses from the "detail" field
in the response_body.