Add global feature flag for file upload
This should make turning the feature on easier - rather than running a
task to enable it for all of the groups, we can turn it on globally with
an environment variable.

This also means that if we need to turn the feature off again, we won't
have to reconfigure any test groups.
modified:   config/settings.yml #	modified:
spec/requests/pages/type_of_answer_controller_spec.rb #