Change query used to update mou_signers_list
We want to add organisation_admins to the MOU updates mailing list, even
if they haven't agreed to the MOU themselves.

When using Mailchimp, we want to be able to target members based on
their role.

This query changes how members for the mou_signers list are generated
and gives each member a role value.