Add task for moving users and groups from one organisation to another
It often happens that organisations are merged into each other. In the
GOV.UK organisation API, this is modelled as an organisation being
closed and superseded by another. When this happens however this doesn't
affect existing users and groups in our system.

We could move the existing users and groups using a combination of the
edit users page for super admins and existing rake tasks, but this would
be time-consuming and error prone.

This commit adds a task to merge a closed `source_organisation` into a
`target_organisation`, so that all users and groups end up in the target
organisation.

We add the restriction that the source organisation should be closed
before the task is run so that no users can be added to that
organisation while the task is running.