Add tests for double click prevention on create a form page
In production we've been seeing 1% to 2% of create a form journeys
result in multiple forms with the same name being created. After
spending time looking at logs and experimenting I've concluded that is
likely because the browser is sometimes sending multiple requests if the
button to submit the form is repeatedly activated.

We want to make sure that on the create a form page clicking the submit
button more than once (or using the enter key more than once) doesn't
result in multiple forms being created.

This commit adds an integration test using a browser to test this.

Normally we have the `prevent-double-click` JavaScript from GOV.UK
Frontend to stop the form being submitted more than once. I haven't been
able to find any bug in this, or recreate the issue of multiple requests
being sent except by having JavaScript disabled, so I've also added a
test which has JavaScript disabled.

This test is currently failing, because we don't have anything to
prevent a double click except by JavaScript. Subsequent commits will
turn this test green.

To be able to disable JavaScript in the test browser I needed to use a
different Capybara driver as there is no easy way to disable JavaScript
with Selenium and ChromeDriver. [Ferrum] via [Cuprite] does include an
easy way to disable JavaScript, and is also very straightforward to
install as it is pure Ruby and works with Capybara the same as Selenium.
These are fairly new gems, but appear to be well supported. There are
also some bugs, however, including the double click not working which I
needed to workaround.

Annoyingly, we have to have both Selenium/ChromeDriver and
Cuprite/Ferrum, as our axe-core accessibility tests don't yet work with
the latter, and as mentioned above Selenium won't let us easily disable
JavaScript.

[Ferrum]: https://github.com/rubycdp/ferrum
[Cuprite]: https://github.com/rubycdp/cuprite