Add server-side prevent double click for create a form
We want to try and make sure that if a user agent sends two create form
requests at or nearly at the same time, we don't end up with forms that
they didn't mean to create. We have the prevent-double-click JavaScript
from GOV.UK Frontend to prevent most cases, but JavaScript is not 100%
reliable, so we also need a server-side solution.

The typical way to make a create request idempotent is to have a
uniqueness constraint in the database.

I experimented with locks of various kinds, but in the end I couldn't
find something as reliable.

Because we don't keep the form records in the forms-admin database yet,
this commit adds a new `create_form_events` table that we can enforce
our constraint on. This can be thought of as a sort of event coalescing.

However, we don't actually want to restrict what forms can be named, so
instead we add `dedup_version` column, which we can increment if there
is a name collision that doesn't look to be due to a double click. This
is partly inspired by the optimistic locking in ActiveRecord.