Add checking end to end tests to PR review checklist
We occasionally have cases where the end to end tests need to be
changed to deploy changes because they rely on content in the app. It
helps to think about this before merging a PR, so this commit adds a
reminder to the PR review checklist to think about it.