Add service to build features report using live form data
Add a service that uses the FormDocuments for live forms retrieved
from forms-api to construct the features usage report. The
FormDocument is the snapshot of the form made when a form is made
live. Previously, we were using the data for the draft version of the
form for live forms which can be different if there have been edits to
the draft since the form was last made live.