Use new feature_report_service for features report
Replace using the response from the forms-api /api/v1/reports/features
with using the feature_report_service, which gets the form documents
from forms-api for all live forms and constructs the report within
forms-admin. This means that only data from live forms, and not
changes in draft versions of live forms will be used to construct the
report.