Allow super admin users to delete groups with no forms
This commit adds routes and actions for deleting a group, including a
confirmation page.

The link to delete a group only shows if the group does not have forms
in it. If the group somehow does have forms in it when the delete is
confirmed an error is shown to the user on the confirmation page (and
the group is not deleted). This whole bit probably needs polish.

For now only super admins are permitted to delete groups.