Add report to list questions with a given answer type
Add a method that will return all questions for a given answer type
by retrieving all live questions from forms-api and filtering them.

Under the hood, this uses pagination to ensure we're not retrieving
and passing all questions in a single database query and API response.
We may experience a degridation in peformance as more forms are added
to the platform, but we can think of ways to optimise later if it
becomes an issue.