Add route to list questions by answer type
Add view/controller for listing all live questions with a given answer
type. This view contains a link to download the filtered questions as
a CSV.

Link to this new report from the the numbers in the "Answer type
usage" table on the "Feature and answer type usage in live forms"
report.