Add report route to list questions with add another answer
Add a report route that lists all questions in live forms with add
another answer enabled. Link to this from the feature usage report
page.

We have an existing separate report for "All forms with add another
answer" which lists questions forms in all states, not just live
forms. I have kept this report as it is as this may be of use.