Add exit pages controller
Adds a new and create endpoint for exit pages - as well as a new view. These make use of the exit page input object, and will have the effect of creating a condition that is formatted as an exit page - this means it has no goto_page_id set, and must instead have exit_page_heading and exit_page_markdown set.