Add exit page input object
Adds an input object in preparation for the new exit page endpoint. This will allow the submission for a form containing exit page heading and markdown data, as well as carried over data for answer values relating to the route. This answer value is set from the previous new route page, so has been passed in when the input object is initialised