Prevent multiple conditions created on single page
Currently, form creators can create multiple conditions on a single
page, by submitting the create form multiple times.

This commit prevents this by checking that the condition does not
already exist before creating it.

Secondary skips can still be created because they are not stored in the
page's routing conditions.