Log all validation errors on input objects
Add an after_validation callback that will add a validation_errors
logging attribute if there were any errors added by the validation.

Log the attribute name and the type of the attribute, rather than the
actual error message. The reasoning for this is that we are probably
less likely to rename the attributes and error types than we are to
change the content of the error message.