Use method to get page once
We have a method in `pages_controller` to either use the page we have
already retrieved from the PageRepository, else retreive it. The
delete and destroy controller actions weren't using this method
because they had been moved from a separate controller and the
implementations weren't updated to use this. Use the method so that we
only ever need to retreive the page once per request.