Log the answer_type when creating or editing questions
Add an answer_type attribute to log lines for pages where we are
adding or editing questions. Get the answer_type from the
DraftQuestion we are editing if there is one present, else from the
page. Add this as an `after_action` in the controller so that we make
sure we log the answer_type on requests that set it as part of the
controller action.