Fix server error when question with route is at the end of a form
We’re seeing errors in production because on the routes page we want to
tell the user about the any other answer route, but in the case where
the question with the route has been moved to the end of the form we
can't find the question object for the next question number.

This commit fixes the error by not trying to find the next question if
the route question is at the end of the form, and instead showing
content telling the user the form filler will see the check your answers
page next.

Curiously it looks like something like this was originally the plan,
because there was already a spec for the show routes view for a question
at the end of the form; I just fluffed the test when writing it 😅