Use new CloudWatch metrics
We've now been sending modified metrics to Cloudwatch for the form
started and submitted event for over a month. We can now use these
new metrics that have a different name, namespace and dimensions for
displaying the form metrics to form creators. After this, we can stop
sending the old metrics from forms-runner.