Fix deleting group when forms have been created in it
We were getting foreign key constraint errors when trying to delete
groups because the create_form_event table has associations with groups.

We currently only use the events for checking whether a form has already
been created with the same name (to catch double clicks), so we can just
delete the events.

This commit fixes the errors by setting `ON DELETE CASCADE` to the
foreign key constraint (thanks @thomasiles for the suggestion). The
migration to change the constraint is split into two to make sure that
the validation does not lock any tables, as recommended by [[1], [2]].

We also make the same change to the users foreign key in the create form
events table, just to avoid possibly hitting the same issue again in
future.

[1]: https://github.com/ankane/strong_migrations?tab=readme-ov-file#adding-a-foreign-key
[2]: https://www.postgresql.org/docs/17/sql-altertable.html#SQL-ALTERTABLE-NOTES