Change CsvReportsService to receive form_documents
Rather than call FormDocumentsService inside CsvReportsService and
couple those two classes together, change CsvReportService to expect to
be passed a list of form_documents. This allows us to reuse the
filtering logic in FeatureReportService instead of reimplementing it in
CsvReportsService, lets us change the source of documents in future, and
also lets us increases the commonality in the controller actions for
rendering a report in HTML and CSV so they might be combined.