Add sleep after updating user organisation in tests
Through experimentation I found that tests in
spec/features/mou/upgrade_user_changes_mou_spec.rb would often fail
because we were attempting to visit the MOU page too soon after
updating a user's organisation. I assume this is a race condition
between the tests and the redirect to the users page after successfully
updating a user's organisation.

For now I've just added a 1s sleep before visiting the MOU index in the
tests. This could be more graceful (wait to make sure we hit the users
page first?) but this is probably ok for now.