Create volumes for whole directories in the Dockerfile
Previously we created 1 volume for each directory that the application needed
to write to, because we found it didn't have permission to create
subdirectories of a volume.

This is because we weren't doing it correctly. The correct way to do it is to
create the directory in the Dockerfile and set its permissions and owner, then
expose it as a volume.

The "/app/db" directory is a special case here because it already exists with
in the "/app" directory, and maps to the "./db/" directory in the repository.
AWS ECS will take care of copying files existing in that directory to new
volume [1] when mounting it.

[1] https://docs.aws.amazon.com/AmazonECS/latest/developerguide/bind-mounts.html