Change questions_with_answer_type page download CSV link to go to self
Rather than going to the reports#live_csv_questions action for the CSV
of questions with answer type, we should use the same pattern as for the
other feature reports and use the same URL but with a `.csv` appended
for the CSV download link on the questions with answer type report page.

This commit updates the #questions_with_answer_type action to accept
format `csv` as well as `html`, same as the other feature report
actions, and uses that in the view template. We can then also simplify
the live_questions_csv action.

This probably should have been done in PR #1976, but it was missed.