Fix feature report template error when records is empty list
The `report_table` helper can't handle it if `records` is an empty
list and raises an exception; this commit fixes that by not calling it
if `records` is empty. For friendliness, we instead show the user a
message making it clear there are no forms/questions to show. For
completeness we also make the same change to the
questions_with_answer_type report view, even though it's not strictly
necessary.