Fix feature report view specs
These tests have not been working correctly, a bug in how they are
written means they will always pass. In the context of view specs,
`within` is not a Capybara method, but an RSpec matcher.

The simplest way to rewrite these is to use the Capybara `have_table`
matcher.

This commit also rewords some incorrect spec descriptions.