Fix links for draft forms in reports
Previously we were using the live form pages path for all forms in
feature reports tables, but for draft forms this is a 404, and forms
which are live with draft this might have the wrong information.

Instead we should link to the form pages path.

This commit fixes things by making the reports helpers choose the path
for the link href based on the form document tag.