Fix Report model
When `ActiveResource::Base` encounters nested attributes in a JSON
response, it tries to instantiate a class with the classified name of
the parent attribute; i.e. when it sees `form: {...}` in a report
response it tries to instantiate a `Form` object. When we didn't have a
Form class this was fine (because ActiveResource falls back to creating
a new class and then instantiating that), but now we have added the
`Form` model it's trying to use that class , which doesn't work because
that class is not designed for use with ActiveResource.

Instead we create another `Form` class namespaced to `Report`,
which will shadow the Form model and work as expected.