Add #database_attributes method to forms resources
As part of replicating data for forms models we need to be able to
identify which attributes in a response can be stored in the database
and which were generated by forms-api.

This commit adds a `#database_attributes` method to each relevant
API model which returns a hash of only the attributes that have a
database column in the corresponding database model.