Find user by email using case insensitive search
On the User model, we have a constraint `uniqueness: { case_sensitive:
false }` which treats emails case insensitively. However, we are
looking up users using the email from the authentication provider case
sensitively.

This inconsistency led to an incident on 12 Dec 2024 where a user had
two user accounts with different cases - i.e. a.user@example.gov.uk
and a.User@example.gov.uk, and this meant that when they tried to log
in they received an error and could not proceed.

We want to prevent things like this happening in future. We decided
that the solution was to always treat the email as case insensitive.
While email addresses are technically case-senstive, we decided that
this was the best solution because:

Auth0 already lowercases most emails by default (so the email
addresses that we recieve are already being treated as case
insensitive in some places)

We think it is very likely that most government email accounts are
already case insensitive, as this is how common email providers treat
email addresses

We think this matches how most users perceive email addresses

We already treat emails as case-insensitive in other places (for
example, when adding users to groups).